Add support for flashing to the interal eMMC
By default, the Kontron imx8mm's boot is set to boot from SD first
and fall back to SPI NOR flash. To allow booting without having to
change any of the fuses, we flash the bootloader to the SPI NOR
flash in the resin-init-flasher-board script. To be able to do so,
the mtd-utils are added to the build. They're also added to the
regular image, because we'll need them to update the bootloader
once we implement the update process.