Reinstate all missing os-release fields
These fields are read by Balena when registering a device in the cloud.
Balena depends on the default os-release.bb recipe which adds these
fields. Balena then appens their own fields to the list. However, the
Kontron BSP overrides the list of fields to include. Therefor, we
reinstate the missing ones in our meta-balena-kontron layer.